@@ -28,6 +28,7 @@
   "homepage": "https://github.com/yeojz/data-calendar",
   "devDependencies": {
     "autoprefixer": "^5.1.0",
+    "babelify": "^5.0.4",
     "browserify": "^9.0.3",
     "clean-css": "^3.1.4",
     "node-sass": "^2.0.1",
@@ -35,6 +36,6 @@
   },
   "dependencies": {
     "moment": "^2.9.0",
-    "react": "^0.12.2"
+    "react": "^0.13.1"
   }
 }
